delete authorization

+++ App/controllers/ListingController.php
@@ -2,7 +2,9 @@
 
 namespace App\Controllers;
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
+use Framework\Authorization;
 
 class ListingController {
 
@@ -20,7 +22,7 @@ class ListingController {
      * @return void
      */
     public function index() {
-        $listings  = $this->db -> query('SELECT * FROM listings') -> fetchAll();
+        $listings  = $this->db -> query('SELECT * FROM listings ORDER BY created_at DESC') -> fetchAll();
         loadView('listings/index', ['listings' => $listings]);
         
     }
@@ -71,7 +73,7 @@ class ListingController {
   
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -128,6 +130,16 @@ class ListingController {
             ErrorController::notFound('Listing not found');
             return;
         }
+
+        //authorization
+        if(!Authorization::isOwner($listing->user_id)){
+            $_SESSION['error_message'] =  'You are not allowed to delete this listing';
+            header('Location: /public/listings/' . $listing->id);
+            exit;
+        }
+        
+        
+
         
         $this -> db -> query('DELETE FROM listings WHERE id = :id', $params);
